[Refactor] Made the tilt motor its own subsystem
